# openapi.json
{
  "components": {
    "schemas": {
...
      "ExecutionResult": {
        "additionalProperties": false,
        "description": "Truthful terminal report of an execution attempt (BRD §23).",
        "properties": {
          "actual_cost": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Actual Cost"
          },
          "broker_order_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Broker Order Id"
          },
          "completed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "cycle_id": {
            "title": "Cycle Id",
            "type": "string"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "failed_legs": {
            "items": {
              "$ref": "#/components/schemas/FailedLeg"
            },
            "title": "Failed Legs",
            "type": "array"
          },
          "filled_legs": {
            "items": {
              "$ref": "#/components/schemas/FilledLeg"
            },
            "title": "Filled Legs",
            "type": "array"
          },
          "order_ids": {
            "items": {
              "type": "string"
            },
            "title": "Order Ids",
            "type": "array"
          },
          "recovery_action": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "the legging-risk recovery step taken for a non-clean fill (e.g. CANCEL_UNFILLED_LEGS, UNWIND_FILLED_LEGS) — P5-BE-13",
            "title": "Recovery Action"
          },
          "slippage": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Slippage"
          },
          "status": {
            "$ref": "#/components/schemas/ExecutionStatus"
          },
          "submitted_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Submitted At"
          }
        },
        "required": [
          "cycle_id",
          "status"
        ],
        "title": "ExecutionResult",
        "type": "object"
      },
      "ExecutionStatus": {
        "description": "Terminal state of an execution attempt (BRD §23).",
        "enum": [
          "FILLED",
          "PARTIALLY_FILLED",
          "FAILED",
          "CANCELLED"
        ],
        "title": "ExecutionStatus",
        "type": "string"
      },
      "FailedLeg": {
        "additionalProperties": false,
        "description": "A leg that did not fill, with the reason.",
        "properties": {
          "leg_symbol": {
            "title": "Leg Symbol",
            "type": "string"
          },
          "reason": {
            "title": "Reason",
            "type": "string"
          }
        },
        "required": [
          "leg_symbol",
          "reason"
        ],
        "title": "FailedLeg",
        "type": "object"
      },
...
      "FillOut": {
        "properties": {
          "filled_at": {
            "format": "date-time",
            "title": "Filled At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "leg_symbol": {
            "title": "Leg Symbol",
            "type": "string"
          },
          "order_id": {
            "title": "Order Id",
            "type": "integer"
          },
          "price": {
            "title": "Price",
            "type": "number"
          },
          "qty": {
            "title": "Qty",
            "type": "number"
          },
          "slippage": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Slippage"
          }
        },
        "required": [
          "id",
          "order_id",
          "leg_symbol",
          "qty",
          "price",
          "filled_at"
        ],
        "title": "FillOut",
        "type": "object"
      },
      "FilledLeg": {
        "additionalProperties": false,
        "description": "A leg that filled, with realized price and slippage vs the expected price.",
        "properties": {
          "filled_at": {
            "format": "date-time",
            "title": "Filled At",
            "type": "string"
          },
          "leg_symbol": {
            "title": "Leg Symbol",
            "type": "string"
          },
          "price": {
            "minimum": 0.0,
            "title": "Price",
            "type": "number"
          },
          "qty": {
            "exclusiveMinimum": 0.0,
            "title": "Qty",
            "type": "integer"
          },
          "slippage": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Slippage"
          }
        },
        "required": [
          "leg_symbol",
          "qty",
          "price",
          "filled_at"
        ],
        "title": "FilledLeg",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "OrderOut": {
        "properties": {
          "broker_order_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Broker Order Id"
          },
          "cycle_id": {
            "title": "Cycle Id",
            "type": "string"
          },
          "fills": {
            "items": {
              "$ref": "#/components/schemas/FillOut"
            },
            "title": "Fills",
            "type": "array"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "legs": {
            "anyOf": [
              {
                "items": {},
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Legs"
          },
          "order_class": {
            "title": "Order Class",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "submitted_at": {
            "format": "date-time",
            "title": "Submitted At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "cycle_id",
          "order_class",
          "status",
          "submitted_at",
          "fills"
        ],
        "title": "OrderOut",
        "type": "object"
      },
...
      "RiskCheckOut": {
        "properties": {
          "checks": {
            "anyOf": [
              {
                "items": {},
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Checks"
          },
          "cycle_id": {
            "title": "Cycle Id",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "modifications": {
            "anyOf": [
              {
                "items": {},
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Modifications"
          },
          "verdict": {
            "title": "Verdict",
            "type": "string"
          },
          "violations": {
            "anyOf": [
              {
                "items": {},
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Violations"
          },
          "warnings": {
            "anyOf": [
              {
                "items": {},
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Warnings"
          }
        },
        "required": [
          "id",
          "cycle_id",
          "verdict"
        ],
        "title": "RiskCheckOut",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Autonomous Adaptive Portfolio Hedge Agent",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/execution": {
      "get": {
        "description": "The most recently persisted order for a cycle, as an ExecutionResult.\n\ncycle_id is optional; omitting it returns the most recently submitted\norder across all cycles (mirrors ``GET /workflow-state``'s \"omit for\nlatest\" convention). Raises 404 when no order has been persisted for that\ncycle — a NO_HEDGE decision never reaches ``POST /execute``, so having no\norder to report is a legitimate outcome, not a server error.",
        "operationId": "execution_result_execution_get",
        "parameters": [
          {
            "description": "Cycle to inspect; omit for latest",
            "in": "query",
            "name": "cycle_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Cycle to inspect; omit for latest",
              "title": "Cycle Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExecutionResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Execution Result",
        "tags": [
          "exec-readback"
        ]
      }
    },
...
    "/orders": {
      "get": {
        "description": "Orders in insertion order with their fills nested, optionally by cycle.",
        "operationId": "orders_orders_get",
        "parameters": [
          {
            "description": "restrict to one execution cycle",
            "in": "query",
            "name": "cycle_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "restrict to one execution cycle",
              "title": "Cycle Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/OrderOut"
                  },
                  "title": "Response Orders Orders Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Orders",
        "tags": [
          "exec-readback"
        ]
      }
    },
...
    "/risk/checks": {
      "get": {
        "description": "Risk checks in insertion order, optionally filtered to one cycle.",
        "operationId": "risk_checks_risk_checks_get",
        "parameters": [
          {
            "description": "restrict to one risk-evaluation cycle",
            "in": "query",
            "name": "cycle_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "restrict to one risk-evaluation cycle",
              "title": "Cycle Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RiskCheckOut"
                  },
                  "title": "Response Risk Checks Risk Checks Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Risk Checks",
        "tags": [
          "exec-readback"
        ]
      }
    },
...
  }
}